pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "scout"
version = "0.1.0"
description = "Scraping Careers & Opportunities Using Technology (SCOUT) - A job listing scraper system"
authors = [
  { name = "Sean Stafford" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
]
requires-python = ">=3.9"

dependencies = [
    # Web scraping
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "markdownify>=0.11.0",

    # Data processing
    "pandas>=2.0.0",
    "numpy>=1.24.0",

    # Database
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",

    # CLI
    "typer>=0.9.0",

    # Utilities
    "tqdm>=4.65.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ipykernel>=6.25.0",
    "jupyterlab>=4.0.0",
]

# Future: for LLM-based job filtering/analysis
analysis = [
    "openai>=1.0.0",
    "anthropic>=0.8.0",
]


[tool.ruff]
# Exclude commonly ignored directories
exclude = [
    ".cache",
    ".git",
    ".ipynb_checkpoints",
    ".pytest_cache",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "__pycache__",
    "build",
    "data",
    "dist",
    "notebooks",
    "outs",
    "transplanted_files",
    "venv",
]

line-length = 100
indent-width = 4
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E731", "E501"]  # E731: lambda assignment, E501: line too long
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"

# Use spaces for indentation
indent-style = "space"

# Respect magic trailing commas
skip-magic-trailing-comma = false

# Auto-detect line endings
line-ending = "auto"

# Format code in docstrings
docstring-code-format = true
docstring-code-line-length = "dynamic"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short --strict-markers -m 'not slow' --disable-warnings"
markers = [
    "unit: Unit tests (fast, mocked dependencies)",
    "integration: Integration tests (slower, may use real database/network)",
    "slow: Tests that take more than 1 second",
    "scraper: Tests for scraper functionality",
    "database: Tests that require PostgreSQL",
]

[tool.coverage.run]
source = ["scout"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/_internal.py",
]
